Bugs
- Notion dynamic file upload
- Notion document metadata

+++ ift_rag/assets/notion.py
@@ -5,9 +5,9 @@ from ..utils import notion_parser
 from llama_index.core import Document
 import pandas as pd
 import traceback
-import time
 import json
 import os
+import datetime
 
 @dg.asset(
     kinds=["Notion"],
@@ -31,6 +31,7 @@ def notion_page_ids(context: dg.AssetExecutionContext, notion: Notion) -> dg.Out
             ]),
             "created_time": page["created_time"],
             "last_edited_time": page["last_edited_time"],
+            "last_edited_by": page["last_edited_by"]["id"],
             "url": page["url"],
             "archived": page["archived"],
         }
@@ -76,17 +77,28 @@ def notion_page_data(context: dg.AssetExecutionContext, page_ids: pd.DataFrame, 
     downloaded_ids = pd.Series(os.listdir(config.local_path)).str.replace(".json", "").to_list()
 
     query = ~page_ids["id"].isin(downloaded_ids)
-    ids = page_ids.loc[query, "id"].to_list()
-    context.log.info(f"To do: {len(ids)} IDs")
+    pages = page_ids.loc[query].reset_index(drop=True).copy().to_dict("records")
 
-    for index, id in enumerate(ids):
+    context.log.info(f"To do: {len(pages)} IDs")
+
+    for index, page in enumerate(pages):
+
+        page = {
+            key: value if "time" not in key else str(value)
+            for key, value in page.items()
+        }
         try:
-            data = notion.get_all_blocks(id)
-            context.log.info(f"{index}) Extracted data for {id}")
+            data = notion.get_all_blocks(page['id'])
+            context.log.info(f"{index}) Extracted data for {page['id']}")
 
-            file_path = os.path.join(config.local_path, f"{id}.json")
+            file_path = os.path.join(config.local_path, f"{page['id']}.json")
+            
+            json_data = {
+                "metadata": page,
+                "data": data
+            }
             with open(file_path, "w") as file:
-                json.dump(data, file, indent=4)
+                json.dump(json_data, file, indent=4)
 
             context.log.info(f"Created {file_path}")
             metadata["downloads"] += 1
@@ -94,8 +106,6 @@ def notion_page_data(context: dg.AssetExecutionContext, page_ids: pd.DataFrame, 
             context.log.warning(f"There was an error with page Notion page ID - {id}\n\n{traceback.format_exc()}")
             metadata["failed"] += 1
     
-        time.sleep(1)
-
     return dg.MaterializeResult(metadata=metadata)
 
 
@@ -127,8 +137,16 @@ def notion_markdown_documents(context: dg.AssetExecutionContext, config: NotionB
         markdown = []
         
         with open(file_path, "r") as file:
-            data: list[dict] = json.load(file)
+            json_data: dict = json.load(file)
         
+        data: list[dict] = json_data.get("data", [])
+        metadata: dict = json_data.get("metadata", {})
+
+        metadata = {
+            key: datetime.datetime.strptime(value, "%Y-%m-%d %H:%M:%S") if "time" in key else value 
+            for key, value in metadata.items()
+        }
+
         for index, document in enumerate(data):
 
             if document["type"] in config.skip_block_types:
@@ -140,13 +158,18 @@ def notion_markdown_documents(context: dg.AssetExecutionContext, config: NotionB
             if len(block.children) != 0 and not isinstance(block, notion_parser.Table):
                 markdown.append(notion_parser.get_child_markdown(block, ""))
 
+        if not markdown:
+            context.log.info(f"No Markdown for {file_path}")
+            os.remove(file_path)
+            continue
+        
         archive_path = add_slash(config.archive_json_path) + file_name
         params = {
             "text": "".join(markdown),
             "metadata": {
                 "path": archive_path,
-                "page_id": page_id,
-                "source": "notion"
+                "source": "notion",
+                **metadata
             }
         }
 
@@ -157,8 +180,10 @@ def notion_markdown_documents(context: dg.AssetExecutionContext, config: NotionB
         context.log.info(f"Uploaded {page_id} Markdown to {path}")
 
         minio.upload(data, archive_path)
-        os.remove(file_path)
-        context.log.info(f"Archived file {page_id} - {archive_path}")
+
+        if not config.debug:
+            os.remove(file_path)
+            context.log.info(f"Archived file {page_id} - {archive_path}")
         
 
     metadata = {
